dsavisual/requirements.txt
django>=5.0,<7.0
djangorestframework>=3.15
django-cors-headers>=4.3
markdown>=3.5
RestrictedPython>=7.0
